+++ packages/client/wallets/aa/package.json
@@ -4,13 +4,11 @@
         "@ambire/signature-validator": "1.3.1",
         "@crossmint/common-sdk-base": "0.0.3",
         "@datadog/browser-logs": "4.42.2",
-        "@fireblocks/ncw-js-sdk": "10.0.1",
-        "@web3auth/base": "^6.0.1-alpha.0",
+        "@web3auth/base": "^7.3.1",
         "@web3auth/ethereum-provider": "^7.3.1",
         "@web3auth/single-factor-auth": "^7.3.0",
         "@zerodev/ecdsa-validator": "5.0.0",
         "@zerodev/session-key": "5.0.0",
-        "@zerodev/sdk-legacy": "npm:@zerodev/sdk@4.5.2",
         "@zerodev/sdk": "5.0.0",
         "@zerodev/presets": "5.0.0",
         "permissionless": "0.0.31",
