+++ packages/client/wallets/aa/src/CrossmintAASDK.test.ts
@@ -28,50 +28,4 @@ describe("CrossmintAASDK", () => {
         });
     });
     const allChains = [...EVM_CHAINS, ...BLOCKCHAIN_TEST_NET];
-
-    describe.each(allChains)("getOrCreateWallet - With Fireblocks signer and %s network", (chain) => {
-        const mockUser: UserIdentifier = {
-            type: "email",
-            email: "test@example.com",
-        };
-        const mockWalletConfig: WalletConfig = {
-            signer: {
-                type: "FIREBLOCKS_NCW",
-                passphrase: "1234",
-            },
-        };
-
-        it("should attempt to create or get a wallet", async () => {
-            const spy = jest.spyOn(sdk, "getOrCreateWallet");
-
-            try {
-                await sdk.getOrCreateWallet(mockUser, chain, mockWalletConfig);
-            } catch (e) {}
-
-            expect(spy).toHaveBeenCalledWith(mockUser, chain, mockWalletConfig);
-            spy.mockRestore();
-        });
-    });
-
-    describe.each(allChains)("getOrCreateWallet - With ethers.Signer and %s network", (chain) => {
-        const mockUser: UserIdentifier = {
-            type: "email",
-            email: "test@example.com",
-        };
-        const randomWallet = ethers.Wallet.createRandom();
-        const mockWalletConfig: WalletConfig = {
-            signer: randomWallet,
-        };
-
-        it("should attempt to create or get a wallet", async () => {
-            const spy = jest.spyOn(sdk, "getOrCreateWallet");
-
-            try {
-                await sdk.getOrCreateWallet(mockUser, chain, mockWalletConfig);
-            } catch (e) {}
-
-            expect(spy).toHaveBeenCalledWith(mockUser, chain, mockWalletConfig);
-            spy.mockRestore();
-        });
-    });
 });
