+++ packages/client/wallets/aa/src/CrossmintAASDK.ts
@@ -1,20 +1,12 @@
 import { CrossmintService } from "@/api";
-import { EVMAAWallet, getChainIdByBlockchain, getZeroDevProjectIdByBlockchain, isEVMBlockchain } from "@/blockchain";
-import type { CrossmintAASDKInitParams, SignerType, WalletConfig } from "@/types";
-import { CURRENT_VERSION, WalletSdkError, ZERO_DEV_TYPE, createOwnerSigner, errorToJSON } from "@/utils";
-import type { SmartAccountSigner } from "@alchemy/aa-core";
+import { EVMAAWallet, getChainIdByBlockchain, isEVMBlockchain } from "@/blockchain";
+import type { CrossmintAASDKInitParams, WalletConfig } from "@/types";
+import { CURRENT_VERSION, WalletSdkError, ZERO_DEV_TYPE, createSmartAccountSigner, errorToJSON } from "@/utils";
 import { createEcdsaKernelAccountClient } from "@zerodev/presets/zerodev";
-import { ZeroDevEthersProvider } from "@zerodev/sdk-legacy";
 import { polygonMumbai } from "viem/chains";
 
-import {
-    BlockchainIncludingTestnet,
-    EVMBlockchainIncludingTestnet,
-    UserIdentifier,
-    UserIdentifierParams,
-} from "@crossmint/common-sdk-base";
+import { BlockchainIncludingTestnet, UserIdentifier, UserIdentifierParams } from "@crossmint/common-sdk-base";
 
-import { EVMAAWalletViem } from "./blockchain/wallets/EVMAAWalletViem";
 import { logError, logInfo } from "./services/logging";
 import { parseUserIdentifier } from "./utils/user";
 
@@ -32,8 +24,7 @@ export class CrossmintAASDK {
     async getOrCreateWallet<B extends BlockchainIncludingTestnet = BlockchainIncludingTestnet>(
         user: UserIdentifierParams,
         chain: B,
-        walletConfig: WalletConfig,
-        signerType?: SignerType
+        walletConfig: WalletConfig
     ) {
         try {
             logInfo("[GET_OR_CREATE_WALLET] - INIT", {
@@ -46,19 +37,21 @@ export class CrossmintAASDK {
             }
 
             const userIdentifier = parseUserIdentifier(user);
-            const owner = await createOwnerSigner(
-                userIdentifier,
-                chain,
-                walletConfig,
-                this.crossmintService,
-                signerType
-            );
-            let evmAAWallet;
-            if (signerType === "viem") {
-                evmAAWallet = await this.handleViemSigner(userIdentifier, chain, owner);
-            } else {
-                evmAAWallet = await this.handleEthersSigner(userIdentifier, chain, owner);
-            }
+            const smartAccountSigner = await createSmartAccountSigner(chain, walletConfig);
+
+            const kernelClient = await createEcdsaKernelAccountClient({
+                projectId: process.env.ZERODEV_PROJECT_ID!,
+                chain: polygonMumbai,
+                signer: smartAccountSigner!,
+            });
+
+            const address = smartAccountSigner!.address;
+            const evmAAWallet = new EVMAAWallet(kernelClient, chain, this.crossmintService, smartAccountSigner!);
+            const abstractAddress = await evmAAWallet.getAddress();
+            const { sessionKeySignerAddress } = await this.crossmintService.createSessionKey(abstractAddress);
+
+            evmAAWallet.setSessionKeySignerAddress(sessionKeySignerAddress);
+            await this.storeAbstractWallet(userIdentifier, chain, abstractAddress, address, sessionKeySignerAddress);
             logInfo("[GET_OR_CREATE_WALLET] - FINISH", {
                 userEmail: user.email!,
                 chain,
@@ -88,30 +81,6 @@ export class CrossmintAASDK {
         });
     }
 
-    private async initializeEthersSigner<B extends BlockchainIncludingTestnet>(owner: SmartAccountSigner, chain: B) {
-        const zDevProvider = await ZeroDevEthersProvider.init("ECDSA", {
-            projectId: getZeroDevProjectIdByBlockchain(chain),
-            owner,
-            opts: {
-                paymasterConfig: {
-                    policy: "VERIFYING_PAYMASTER",
-                },
-            },
-        });
-
-        return { zDevProvider };
-    }
-
-    private async initializeViemSigner<B extends BlockchainIncludingTestnet>(owner: any, chain: B) {
-        const kernelClient = await createEcdsaKernelAccountClient({
-            projectId: process.env.ZERODEV_PROJECT_ID!,
-            chain: polygonMumbai,
-            signer: owner,
-        });
-
-        return kernelClient;
-    }
-
     private async storeAbstractWallet<B extends BlockchainIncludingTestnet>(
         userIdentifier: UserIdentifier,
         chain: B,
@@ -129,35 +98,5 @@ export class CrossmintAASDK {
             baseLayer: "evm",
             chainId: getChainIdByBlockchain(chain),
         });
-    }
-
-    private async handleViemSigner<B extends EVMBlockchainIncludingTestnet = EVMBlockchainIncludingTestnet>(
-        userIdentifier: UserIdentifier,
-        chain: B,
-        owner: SmartAccountSigner
-    ): Promise<EVMAAWalletViem> {
-        const kernelClient = await this.initializeViemSigner(owner, chain);
-        const address = await owner.getAddress();
-        const evmAAWallet = new EVMAAWalletViem(kernelClient, chain, this.crossmintService, owner);
-        const abstractAddress = await evmAAWallet.getAddress();
-        const { sessionKeySignerAddress } = await this.crossmintService.createSessionKey(abstractAddress);
-        evmAAWallet.setSessionKeySignerAddress(sessionKeySignerAddress);
-        await this.storeAbstractWallet(userIdentifier, chain, abstractAddress, address, sessionKeySignerAddress);
-        return evmAAWallet;
-    }
-
-    private async handleEthersSigner<B extends EVMBlockchainIncludingTestnet = EVMBlockchainIncludingTestnet>(
-        userIdentifier: UserIdentifier,
-        chain: B,
-        owner: SmartAccountSigner
-    ): Promise<EVMAAWallet> {
-        const address = await owner.getAddress();
-        const { zDevProvider } = await this.initializeEthersSigner(owner, chain);
-        const evmAAWallet = new EVMAAWallet(zDevProvider!, this.crossmintService, chain);
-        const abstractAddress = await evmAAWallet.getAddress();
-        const { sessionKeySignerAddress } = await this.crossmintService.createSessionKey(abstractAddress);
-        evmAAWallet.setSessionKeySignerAddress(sessionKeySignerAddress);
-        await this.storeAbstractWallet(userIdentifier, chain, abstractAddress, address, sessionKeySignerAddress);
-        return evmAAWallet;
     }
 }
