+++ packages/client/wallets/aa/src/blockchain/wallets/index.ts
@@ -1,3 +1,1 @@
 export * from "./EVMAAWallet";
-export * from "./EVMAAWalletViem";
-export * from "./FireblocksNCWallet";
