+++ packages/client/wallets/aa/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./api";
 export * from "./blockchain";
-export * from "./storage";
 export * from "./types";
 export * from "./utils";
