--- packages/client/wallets/aa/src/storage/index.ts
@@ -1,2 +1,0 @@
-export * from "./PasswordEncryptedLocalStorage";
-export * from "./LocalStorageRepository";
