+++ packages/client/wallets/aa/src/utils/signer.ts
@@ -1,6 +1,4 @@
-import { CrossmintService } from "@/api";
 import {
-    FireblocksNCWallet,
     getBlockExplorerByBlockchain,
     getChainIdByBlockchain,
     getDisplayNameByBlockchain,
@@ -9,39 +7,17 @@ import {
     getUrlProviderByBlockchain,
     getWeb3AuthBlockchain,
 } from "@/blockchain";
-import { FireblocksNCWSigner, SignerType, UserIdentifier, WalletConfig, Web3AuthSigner } from "@/types";
+import { Provider, WalletConfig, Web3AuthSigner } from "@/types";
 import { parseToken } from "@/utils";
-import type { SmartAccountSigner } from "@alchemy/aa-core";
 import { CHAIN_NAMESPACES } from "@web3auth/base";
 import { EthereumPrivateKeyProvider } from "@web3auth/ethereum-provider";
 import { Web3Auth } from "@web3auth/single-factor-auth";
 import { providerToSmartAccountSigner } from "@zerodev/sdk";
-import { convertEthersSignerToAccountSigner, getRPCProviderOwner } from "@zerodev/sdk-legacy";
-import { Signer } from "ethers";
 
 import { BlockchainIncludingTestnet } from "@crossmint/common-sdk-base";
 
-export async function createOwnerSigner(
-    userIdentifier: UserIdentifier,
-    chain: BlockchainIncludingTestnet,
-    walletConfig: WalletConfig,
-    crossmintService: CrossmintService,
-    signerType?: SignerType
-) {
-    if (isFireblocksNCWSigner(walletConfig.signer)) {
-        let fireblocks: any;
-        if ("walletId" in walletConfig.signer && "deviceId" in walletConfig.signer) {
-            const { passphrase, walletId, deviceId } = walletConfig.signer;
-            fireblocks = await FireblocksNCWallet(userIdentifier, crossmintService, chain, passphrase, {
-                walletId,
-                deviceId,
-            });
-        } else {
-            const { passphrase } = walletConfig.signer;
-            fireblocks = await FireblocksNCWallet(userIdentifier, crossmintService, chain, passphrase, undefined);
-        }
-        return fireblocks.owner;
-    } else if (isWeb3AuthSigner(walletConfig.signer)) {
+export async function createSmartAccountSigner(chain: BlockchainIncludingTestnet, walletConfig: WalletConfig) {
+    if (isWeb3AuthSigner(walletConfig.signer)) {
         const signer = walletConfig.signer;
         const chainId = getChainIdByBlockchain(chain);
         const chainConfig = {
@@ -68,18 +44,15 @@ export async function createOwnerSigner(
             verifierId: sub,
             idToken: signer.jwt,
         });
-        if (signerType === "viem") {
-            return await providerToSmartAccountSigner(provider!);
-        } else {
-            return getRPCProviderOwner(provider);
-        }
-    } else {
-        return convertEthersSignerToAccountSigner(walletConfig.signer as Signer);
+
+        return providerToSmartAccountSigner(provider!);
+    } else if (isProvider(walletConfig.signer)) {
+        return providerToSmartAccountSigner(walletConfig.signer.provider);
     }
 }
 
-function isFireblocksNCWSigner(signer: any): signer is FireblocksNCWSigner & { walletId: string; deviceId: string } {
-    return signer && signer.type === "FIREBLOCKS_NCW";
+function isProvider(signer: any): signer is Provider {
+    return signer && signer.type === "PROVIDER";
 }
 
 function isWeb3AuthSigner(signer: any): signer is Web3AuthSigner {
